endpoint de payment corregido

+++ src/main/java/org/mija/elbuensaborback/application/dto/response/PreferenceResponseDto.java
@@ -1,7 +1,9 @@
 package org.mija.elbuensaborback.application.dto.response;
 
+import java.math.BigDecimal;
+
 public record PreferenceResponseDto(
         String preferenceId,
         Long pedidoId,
-        Double totalPedido
+        BigDecimal totalPedido
 ) {}
